dodawanie tras v1203213

+++ MEETIT/Models/PointValues.cs
@@ -6,8 +6,9 @@ namespace meetit.Models
     {
         [Key]
         public int idPointValues { get; set; }
+
         public int idPoint { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public DateOnly date { get; set; }
         public TimeOnly time { get; set; }
     }
